refactor(h-main)!: simplify TCB thm format config
Also, this commit introduces the following changes.

(1) Make supplement an environment.

(2) The references in a supplement are in gray text.

(3) Add the following TCB style and align settings.
* white_bg
* white_bg (to replace the typo white_by)

(3) Add the following TCB theorems.
* Principle (principle): Princ. / princ.
* Example (example): Ex. / ex.

BREAKING CHANGES:
The format definition of TCB theorems (tcbtheorem)
changes from \crefformat and \Crefformat to an
all-in-one command with supplement environment detection
\defformat{tcb-theorem-environment-name}
{Crrefformat-prefix}{crrefformat-prefix}.
For example, \defformat{figure}{Fig.}{fig.}.